do not rm if it doesn't exist

+++ PKGBUILD
@@ -28,7 +28,7 @@ package() {
   msg "GIT checkout done or server timeout"
   msg "Starting build..."
 
-  rm -rf ${srcdir}/${_gitname}-build
+  if [[ -e ${srcdir}/${_gitname}-build ]]; then rm -rf ${srcdir}/${_gitname}-build; fi
   cp -r ${srcdir}/${_gitname} ${srcdir}/${_gitname}-build
   cd ${srcdir}/${_gitname}-build
